refactor(coaches): migrate attendance to AppPageLayout
Extract useCoachAttendancePageLogic hook and CoachAttendanceContainer.
Replaces manual <Tabs> with AppPageLayout tabs prop.
